// payment-service/src/domain/useCases/refunds/process-refund.uc.interface.ts
export interface ProcessRefundDto {
    bookingId: string;
    userId: string;
    mentorId: string;
    totalAmount: number;
    refundBreakdown: {
        userAmount: number;
        mentorAmount: number;
        platformAmount: number;
        userPercentage: number;
        mentorPercentage: number;
        platformPercentage: number;
    };
}

export interface IProcessRefundUC {
    execute(dto: ProcessRefundDto): Promise<void>;
}